Move oneSessionManager, to be used in benchmarks
